% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_packages.R
\name{load_package_info}
\alias{load_package_info}
\title{load_package_info}
\usage{
load_package_info(...)
}
\arguments{
\item{package_file}{The name of the file where the packages are saved. Defaults to current_packages.rda in the current directory.}
}
\description{
This function loads a list of packages from a file and installs the packages.
CRAN is checked first, followed by bioconductor.
}
\examples{
load_package_info()
}
\keyword{installed}
\keyword{load}
\keyword{packages}

