% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_packages.R
\name{save_package_info}
\alias{save_package_info}
\title{save_package_info}
\usage{
save_package_info(package_file = "./current_packages.rda")
}
\arguments{
\item{package_file}{The name of the file where the packages are saved. Defaults to current_packages.rda in the current directory.}
}
\description{
This function saves a list of currently installed packages into the current working directory
}
\examples{
save_package_info()
}
\keyword{installed}
\keyword{packages}
\keyword{save}

